chore: moved file base_logger.py
